Fix inconsistencies in typegraph.
- stylistic: reachable.h should use size_t for num_nodes_ and size_.
- problematic: next_binding_id looks like a getter, but it actually mutates next_binding_id_. This should be two different functions.

PiperOrigin-RevId: 332953498

+++ pytype/typegraph/reachable.h
@@ -18,8 +18,8 @@ class ReachabilityAnalyzer {
 
  private:
   std::vector<std::vector<int64_t>> adj_;
-  int num_nodes_;  // == adj_.size() == adj_[0].size()
-  int size_;
+  size_t num_nodes_;  // == adj_.size() == adj_[0].size()
+  size_t size_;
 };
 
 }  // namespace devtools_python_typegraph
